Add files via upload
i have tried a lot to fix the issues but, i am very new to programming and don't know what's wrong and how to fix it. the base of the script was made by ChatGPT and i have adjusted it a few times and finally got it to run but then when i would talk it would crash. and now, once again, it doesn't even run without giving me errors.it also doesnt recognize chatterbot even though it is installed